* magick/api.h: Eliminate requirement to include certain system headers prior to including <magick/api.h>. * utilities/gm.c: Move utility implementation to GMCommand() in command.c

+++ magick/colorspace.c
@@ -959,7 +959,7 @@ MagickExport MagickPassFail RGBTransformImage(Image *image,
   /*
     Dump tables
   */
-  for (i=0; i <= MaxMap; i++)
+  for (i=0; i <= (long) MaxMap; i++)
     {
       printf("%5ld: x_map(%g,%g,%g) y_map(%g,%g,%g) z_map(%g,%g,%g)\n",
              i,
